build: add wallet handler mixin
